requirement.txt
streamlit
google-generativeai
python-dotenv
pandas
PyPDF2
google-auth